server game engine
changing to an authoritative server setup.
game engine almost complete on server.
gamestate can be sent to clients.
still unsure of how to sync client gameloop with server gameloop.
tbd

+++ pserver/src/game/Locatable.java
@@ -8,6 +8,6 @@ package game;
  * ******************************
  **/
 public interface Locatable {
-    int getX();
-    int getY();
+    double getX();
+    double getY();
 }
